+++ core/src/main/java/com/epam/esm/exception/EntityNotDeletedException.java
@@ -6,9 +6,9 @@ import lombok.Getter;
 public class EntityNotDeletedException extends ServiceException {
 
     private final String errorCode = "4004";
-    private final Integer id;
+    private final Long id;
 
-    public EntityNotDeletedException(String message, int id) {
+    public EntityNotDeletedException(String message, long id) {
         super(message);
         this.id = id;
     }
